Move ns parser tests to the right package
