+++ src/test/kotlin/io/hemin/wien/parser/namespace/FeedpressParserTest.kt
@@ -1,4 +1,4 @@
-package io.hemin.wien.parser
+package io.hemin.wien.parser.namespace
 
 import assertk.all
 import assertk.assertThat
@@ -8,7 +8,8 @@ import assertk.assertions.prop
 import io.hemin.wien.builder.fake.podcast.FakePodcastBuilder
 import io.hemin.wien.builder.fake.podcast.FakePodcastFeedpressBuilder
 import io.hemin.wien.dom.XmlRes
-import io.hemin.wien.parser.namespace.FeedpressParser
+import io.hemin.wien.parser.NamespaceParser
+import io.hemin.wien.parser.NamespaceParserTest
 import org.junit.jupiter.api.Test
 import org.w3c.dom.Node
