+++ src/test/kotlin/io/hemin/wien/parser/namespace/PodloveSimpleChapterParserTest.kt
@@ -1,4 +1,4 @@
-package io.hemin.wien.parser
+package io.hemin.wien.parser.namespace
 
 import assertk.assertThat
 import assertk.assertions.containsExactly
@@ -9,7 +9,7 @@ import io.hemin.wien.builder.fake.episode.FakeEpisodePodloveBuilder
 import io.hemin.wien.builder.fake.episode.FakeEpisodePodloveSimpleChapterBuilder
 import io.hemin.wien.dom.XmlRes
 import io.hemin.wien.noneHasEnoughDataToBuild
-import io.hemin.wien.parser.namespace.PodloveSimpleChapterParser
+import io.hemin.wien.parser.NamespaceParserTest
 import org.junit.jupiter.api.Test
 import org.w3c.dom.Node
